Added unit testing for actuator block commands
Added buttons and a paragraph to html that call buildCommandBlock with
each hexagon's id to test bit math in the function. Buttons and text are
generated only when the test is being run.

Created test file with a test that loops through each actuator and
compares the hard coded ground truth command to the command generated
from the bit math in the buildCommandBlock function. Ground truth
available for all 128 actuators. Commands are binary strings.

Add babel and jest configs to setup test environment.

To run test: npm run test