Switched to using svelte-moveable
svelte-moveable replaces math that changes hexagon layout based on
orientation, and allows the the container with the hexagon svg to
rotate.

Math inside handleTouchMove is not changed, and touch offset values stay
 relative to the rotated div container.

Container can be rotated with cursor or by button which calls a method
of Moveable to rotate by a number of degrees or rotate to an angle.