Used inline style to set initial orientation of actuators
Added setRotationStyle and setScaleStyle to set the overall style, or
transforms, so the actuators are rotated horizontally or vertically for
the different options. The actuators can still be rotated and reset to
their initial position. Scale is currently intended for overlaying an
inactive array of actuators over the option link button, like to mark the
location on the body(WIP).