Improved device response handling
Errors from the device are handled better, giving the user more feedback in
the message field and preventing actions that should be allowed if an error
is thrown (Ex: showing a device as added when it was not found by antenna)
Message variable was added to the store to make it easier for all components
to access.

Also fixed bug in actuator command. JSON did not like the '00000001'
format so the stings are converted into their decimal forms before being sent