Added support for PreparedSequences
Added components to PreparedSequences page.
Added props to set if a hexagon component is for a preset or not. If it is
for a preset, command is sent for that presetName

Checks mouse movement for sweep preset to trigger sweep in swiped direction.

Added OP_mode to store and act_command to send preset commands.