Changed how style of main view is changed
When option is clicked, sets main_obj to the object with that options
settings and uses those settings to style the main container.

Added div with relative position around the hexagons so they could be
positioned better over background. Could be replaced with draggable feature
in future

Made position of status component dependent on window size.