Added device property to hexagons
Hexagons will now have a dev property, along with id, row, col, which sets
the activeDevice number for the actuator command.
This lets you build larger arrays using the stichables(groups of 7) with
id numbers 0-7, and using dev property to distinguish the different
devices.

This limits the user to hardcoded arrangments of stichables.