Fixed hex array positioning/sizeing
Added transform:translate(-50%,-50%) to effectively set the origin of the
hexagons to the center instead of top left corner.
This allowed the use of more intuitive percentages for positioning the
arrays in the option buttons.
Also helps account for differences in size, since the buttons and arrays
may be different sizes on different screens so 30% with the origin in the
top left corner would be different.
Seems to also let you use the same values for the main container and
the options.

Set app div to center everything so columns line up better.
Adjusted position of status indicator.
Fixed variable name in store.