Fixed bug for tracking touch point location
Touch point location is calculated relative to the bounding box of the
target container. The box was previously set onMount to prevent the
target box being changed when moving over text (id#) in the hexagon
buttons.
This caused problems since the target is moved in virutal touch
page but the box is never updated. Since there is no longer text in the
hexagons, assignment of bounding box is done on touch start.